backend/requirements.txt
# ===========================
# 🔧 Core Django + REST
# ===========================
Django>=4.2,<5.0
djangorestframework>=3.14
django-filter>=23.2
django-cors-headers>=4.0.0
django-environ>=0.11.2
django-extensions>=3.2.3

# ===========================
# 🔐 Authentification
# ===========================
djangorestframework-simplejwt>=5.3.0

# ===========================
# 🗄️ Base de données
# ===========================
psycopg2-binary>=2.9

# ===========================
# 🧠 Deep Learning & Vision
# ===========================
torch>=2.0
torchvision>=0.15
opencv-python-headless>=4.9.0
Pillow>=10.2.0
faiss-cpu>=1.7.4
numpy>=1.25.0
scikit-learn>=1.0
timm>=0.9.2
fashion-clip>=0.1.0
tqdm>=4.66.0

# ===========================
# 📦 Backend utils
# ===========================
python-dotenv>=1.0.1
requests>=2.31.0

# ===========================
# ☁️ Stockage cloud
# ===========================
django-storages>=1.14.2
boto3>=1.34.0

# ===========================
# 🚀 Caching / Celery
# ===========================
django-redis>=5.3.0
redis>=5.0.0
celery>=5.3.4
django-celery-beat>=2.5.0
django-celery-results>=2.5.1